verification/requirements.txt
Flask==2.3.3
Pillow==10.0.1
pytesseract==0.3.10
opencv-python==4.8.1.78
numpy==1.24.3
requests==2.31.0
python-multipart==0.0.6
